+++ R/compute_projection.R
@@ -5,22 +5,33 @@ compute_projection <- function(object) {
     dplyr::do(projection_group(., object$info$iname, object$info$tname,
                                object$info$ename,
                                object$info$k_min, object$info$k_max,
-                               object$info$compute_var,
-                               object$info$only_full_horizon)) |>
+                               object$info$compute_var)) |>
     dplyr::ungroup()
 
-  object$data <- object$data |> dplyr::ungroup()
+  # Individual Level Variance
+  if (object$info$compute_var) {
+    object$data <- object$aggregated |>
+      dplyr::select(object$info$by_est, object$info$ename, zz000k,
+                    zz000mean, zz000varcont) |>
+      dplyr::left_join(object$data,
+                       by = c(object$info$by_est,
+                              object$info$ename,
+                              "zz000k")) |>
+      dplyr::mutate(zz000var = (zz000ytilde - zz000mean)^2 - zz000varcont)
+  } else {
+    object$data <- object$data |>
+      dplyr::mutate(zz000var = NA_real_) # For its convenience in the summary
+  }
 
   return(object)
 }
 
 projection_group <- function(data, iname, tname, ename, k_min, k_max,
-                             compute_var, only_full_horizon) {
+                             compute_var) {
 
   aggregated <- data |>
     dplyr::filter(dplyr::between(zz000k, k_min, k_max)) |>
     dplyr::summarize(zz000mean = stats::weighted.mean(zz000ytilde, w = zz000w),
-                     zz000vartr = weighted.var(zz000ytilde, w = zz000w),
                      zz000w = sum(zz000w),
                      .by = c(ename, "zz000k")) |>
     dplyr::arrange(!!rlang::sym(ename), zz000k)
@@ -29,33 +40,11 @@ projection_group <- function(data, iname, tname, ename, k_min, k_max,
     return(aggregated)
   }
 
-  if (only_full_horizon) {
-    ak_feasible <- aggregated |>
-      dplyr::distinct(!!rlang::sym(ename), zz000k)
-
-    ak_feasible <- ak_feasible |>
-      dplyr::summarize(zz000k_max = max(zz000k), .by = ename) |>
-      dplyr::filter(zz000k_max == k_max) |>
-      dplyr::select(-zz000k_max) |>
-      dplyr::left_join(ak_feasible, by = ename)
-
-    if (nrow(ak_feasible) == 0) {
-      return(aggregated[FALSE, ])
-    }
-
-    aggregated <- ak_feasible |>
-      dplyr::left_join(aggregated, by = c(ename, "zz000k"))
-  }
-
   if (compute_var) {
-    sum_varcont <- ak_feasible |>
-      dplyr::mutate(zz000varcont = purrr::map2_dbl(
-          !!rlang::sym(ename), zz000k,
-          ~ varcont_ek(data, iname, tname, ename, .x, .y)))
-
     aggregated <- aggregated |>
-      dplyr::left_join(sum_varcont, by = c(ename, "zz000k")) |>
-      dplyr::mutate(zz000var = zz000vartr - zz000varcont)
+      dplyr::mutate(zz000varcont = purrr::map2_dbl(
+        !!rlang::sym(ename), zz000k,
+        ~ varcont_ek(data, iname, tname, ename, .x, .y)))
   }
 
   return(aggregated)
@@ -78,7 +67,8 @@ varcont_ek <- function(data, iname, tname, ename, e, k) {
                   !!rlang::sym(tname) == e + k) |>
     dplyr::left_join(v_ek, by = c(iname)) |>
     dplyr::mutate(zz000epsilonhat = zz000ytilde - v_ek) |>
-    dplyr::pull() |>
-    stats::var(na.rm = TRUE) # For unbalanced panels, na.rm = TRUE
-
+    dplyr::summarize(zz000varcont = weighted.var(zz000epsilonhat,
+                                                 w = zz000w,
+                                                 na.rm = TRUE)) |>
+    dplyr::pull(zz000varcont) # Note: for unbalanced panels, na.rm = TRUE
 }